WIP - Save the progress on svg animation
I am saving the progress on the svg animation. I managed to make it blink and I am currently working on the initial animation. CSS transform and animation, though, are leaving the svg in a weird state where it doesn't revert back to scale(1). Looks like it stays slightly off. Going to investigate.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>