Include Bubble anim and use discrete files
Include the bubble animation stuff and move everything to own files, so each animation has two sets of files: one for styles and one for js. These are now included singularly in the html. This is in order to avoid confusion in the developers, so that the final file structure is up to them and their current flow.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>